Update Task Service by adding docs

+++ src/core/use-cases/task-use-cases.ts
@@ -4,7 +4,18 @@ import { Task, TaskCommand, TaskRepository, TaskStatus, TaskUseCases } from '../
 import { CurrentUser } from '../domain/user.js'
 import { TaskPrismaRepository } from '../infrastructure/task-repository.js'
 
+/**
+ * The `TaskService` is a service that is responsable to manage all the `Task` entities on the platform.
+ * This service depends on an abstract `repository` and a user `session` representation.
+ * Each Task has a owner defined by the `userId` property which limits the access to all the users.
+ * A Task can be handled only by the user owner or admin user.
+ */
 export class TaskService implements TaskUseCases {
+  /**
+   * Instance a new TaskService to handle Tasks.
+   * @param taskRepository task repository implementation for persistence.
+   * @param session representation of the user session using this service.
+   */
   constructor(
     private readonly taskRepository: TaskRepository = new TaskPrismaRepository(),
     private readonly session: CurrentUser = new CurrentSession()
@@ -18,7 +29,12 @@ export class TaskService implements TaskUseCases {
     return this.session.user.super ? undefined : this.session.user.id
   }
 
-  async getTask(id: string): Promise<Task> {
+  /**
+   * Get a Task by a given id based on the current user permissions.
+   * @param id Task id.
+   * @returns Existing task or an error when it does not exists or user is not authorized.
+   */
+  public async getTask(id: string): Promise<Task> {
     const task = await this.taskRepository.get(id)
 
     if (!task) {
@@ -32,15 +48,29 @@ export class TaskService implements TaskUseCases {
     return task
   }
 
-  async getAllTasks(): Promise<Task[]> {
+  /**
+   * Get a list of all tasks based on the current user.
+   * @returns Async result containing a list of tasks allowed by the current user.
+   */
+  public async getAllTasks(): Promise<Task[]> {
     return await this.taskRepository.getAll(this.userId)
   }
 
-  async getTasksByStatus(status: TaskStatus): Promise<Task[]> {
+  /**
+   * Get a collection of tasks based on the status.
+   * @param status Status of the Tasks defined by `TaskStatus` enum.
+   * @returns Async result containing a collection of tasks.
+   */
+  public async getTasksByStatus(status: TaskStatus): Promise<Task[]> {
     return await this.taskRepository.getByStatus(status, this.userId)
   }
 
-  async addNewTask(command: TaskCommand): Promise<Task> {
+  /**
+   * Add a new tasks for the current user.
+   * @param command Valida command representing a Task.
+   * @returns Async result containing a new Task.
+   */
+  public async addNewTask(command: TaskCommand): Promise<Task> {
     const task: Omit<Task, 'id'> = {
       createdAt: new Date(),
       status: TaskStatus.Todo,
@@ -53,16 +83,26 @@ export class TaskService implements TaskUseCases {
     return newTask
   }
 
-  async updateExistingTask(id: string, command: TaskCommand): Promise<Task> {
+  /**
+   * Update an existing Task by a given id based on the current user permissions.
+   * @param id Id of the task to be updated.
+   * @param command Valid command representing a Task.
+   * @returns Async result containing the updated Task. This operation can also throw an error based on the current user permissions.
+   */
+  public async updateExistingTask(id: string, command: TaskCommand): Promise<Task> {
     const currentTask = await this.taskRepository.get(id)
+
+    // When the given Task does not exists.
     if (!currentTask) {
       throw new NotFoundError(`There is no Task available for id ${id}.`)
     }
 
+    // When a current user is not the owner of the task.
     if (this.userId && currentTask.userId !== this.userId) {
       throw new ForbiddenError('You do not have permission to change this task.')
     }
 
+    // When current Task is already Archived, it should not be updated.
     if (currentTask.status === TaskStatus.Archived) {
       throw new UnprocessableEntityError("It is not possible to change the information on a Task on 'Archived' status.")
     }
@@ -78,20 +118,31 @@ export class TaskService implements TaskUseCases {
     return updatedTask
   }
 
-  async updateStatus(id: string, status: TaskStatus): Promise<Task> {
+  /**
+   * Update the status of a task. Useful when a user just want to move the Task to a different state. Only owners and admin can change it.
+   * @param id Id of the task.
+   * @param status New state of the task.
+   * @returns Async result containing the updated task. This operation can also throw an error depending on the user permissions.
+   */
+  public async updateStatus(id: string, status: TaskStatus): Promise<Task> {
     const currentTask = await this.taskRepository.get(id)
+
+    // When the given Task does not exists.
     if (!currentTask) {
       throw new NotFoundError(`There is no Task available for id ${id}.`)
     }
 
-    if (this.userId && currentTask.userId !== this.userId) {
+    // Only the user owner or admin user can run this change.
+    if (this.userId && (currentTask.userId !== this.userId || !this.session.user.super)) {
       throw new ForbiddenError('You do not have permission to change this task.')
     }
 
+    // When current Task is already Archived, it should not be updated.
     if (currentTask.status === TaskStatus.Archived) {
       throw new UnprocessableEntityError(`It is not possible to move the Task from 'Archived' status.`)
     }
 
+    // Update the status of the task.
     currentTask.status = status
     const updatedTask = await this.taskRepository.update(id, currentTask)
 
@@ -102,11 +153,22 @@ export class TaskService implements TaskUseCases {
     return updatedTask
   }
 
-  async deleteTask(id: string): Promise<boolean> {
+  /**
+   * Delete an existing Task by the id.
+   * @param id Task Id.
+   * @returns Async result containing if the delete operation succeed. This operation can also throw an error depending on the user permissions.
+   */
+  public async deleteTask(id: string): Promise<boolean> {
     const currentTask = await this.taskRepository.get(id)
     if (!currentTask) {
       throw new NotFoundError(`There is no Task available for id ${id}.`)
     }
+
+    // Only the user owner or admin user can delete a Task.
+    if (this.userId && (currentTask.userId !== this.userId || !this.session.user.super)) {
+      throw new ForbiddenError('You do not have permission to change this task.')
+    }
+
     return await this.taskRepository.delete(id)
   }
 }
